python_backend/requirements.txt
# Core web framework
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
werkzeug==2.3.7

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Data processing
pandas==2.1.0
numpy==1.25.2
scikit-learn==1.3.0

# NLP and text processing
nltk==3.8.1
spacy==3.6.1
PyPDF2==3.0.1
python-docx==0.8.11

# LLM and embeddings
openai==1.10.0  # Updated to be compatible with langchain-openai
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.5
transformers==4.33.1
torch==2.0.1
sentence-transformers==2.2.2
tiktoken>=0.5.2,<0.6.0

# Vector database
chromadb==0.4.18
nomic==2.0.3

# Document generation
pdfkit==1.0.0
markdown==3.5
markdown2==2.4.10

# Utilities
python-dotenv==1.0.0
requests==2.31.0